+++ pkg/resolving/named_service_resolver.go
@@ -7,11 +7,11 @@ import (
 )
 
 // NamedServiceResolverActivatorFunc defines a function for resolving named services.
-type NamedServiceResolverActivatorFunc[T any] func(types.Resolver) func(string) (T, error)
+type NamedServiceResolverActivatorFunc[T any] func(context.Context, types.Resolver) func(string) (T, error)
 
 // CreateNamedServiceResolverActivatorFunc creates a NamedServiceResolverActivatorFunc for resolving named services.
-func CreateNamedServiceResolverActivatorFunc[T any](ctx context.Context) NamedServiceResolverActivatorFunc[T] {
-	return func(resolver types.Resolver) func(string) (T, error) {
+func CreateNamedServiceResolverActivatorFunc[T any]() NamedServiceResolverActivatorFunc[T] {
+	return func(ctx context.Context, resolver types.Resolver) func(string) (T, error) {
 		var nilInstance T
 		requiredServices, _ := ResolveRequiredServices[func() types.NamedService[T]](ctx, resolver)
 		return func(name string) (T, error) {